requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.4
aiosqlite==0.20.0
databases==0.9.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0
pytesseract==0.3.10
PyPDF2==3.0.1
youtube-transcript-api==0.6.2
fpdf2==2.7.9
bcrypt==4.1.3
